+++ src/components/MovieDetailsPage/index.js
@@ -1,0 +1,3 @@
+import MovieDetailsPage from './MovieDetailsPage';
+
+export default MovieDetailsPage;
