+++ src/components/MovieDetailsPage/movieDetailsPage.scss
