+++ src/components/MovieDetailsPageAction/index.js
@@ -1,0 +1,3 @@
+import Action from './Action';
+
+export default Action;
