+++ src/components/MovieDetailsPageFooter/index.js
@@ -1,0 +1,3 @@
+import MovieDetailsPageFooter from './MovieDetailsPageFooter';
+
+export default MovieDetailsPageFooter;
