+++ src/components/MovieDetailsPageFooter/movieDetailsPageFooter.scss
