+++ src/components/MovieDetailsPageNav/MovieDetailsPageNav.jsx
@@ -4,27 +4,18 @@ import { connect } from 'react-redux';
 import { getFilms, getGenres } from '../../modules/module_Search/actions';
 import store from '../../modules/store';
 
-import './style.scss';
+import './movieDetailsPageNav.scss';
 
 
-const Nav = ({ fetchData, store }) =>
-// handler(e) {
-//   e.preventDefault();
-//   console.log(this.props.onSearch);
-// }
-
-  // render() {
-  (
-    <nav className="pageHeader__nav">
-      <div className="logo">films</div>
-      <form action="" method="get" className="nav__form_search">
-        <input type="search" className="nav__search" id="searchInputID" />
-        <button onClick={fetchData} type="submit" className="nav__search_button" />
-      </form>
-    </nav>
-  );
-
-// }
+const Nav = ({ fetchData, store }) => (
+  <nav className="pageHeader__nav">
+    <div className="logo">films</div>
+    <form action="" method="get" className="nav__form_search">
+      <input type="search" className="nav__search" id="searchInputID" />
+      <button onClick={fetchData} type="submit" className="nav__search_button" />
+    </form>
+  </nav>
+);
 
 const mapStateToProps = state => ({ store: state });
 const mapDispatchToProps = dispatch => ({
