+++ src/components/MovieDetailsPageNav/index.js
@@ -1,0 +1,3 @@
+import MovieDetailsPageNav from './MovieDetailsPageNav';
+
+export default MovieDetailsPageNav;
