+++ src/components/MovieDetailsPageNav/movieDetailsPageNav.scss
