+++ src/components/MovieDetailsPageStarRating/StarRating.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './style.scss';
+import './starRating.scss';
 
 const StarRating = ({
   title, duration, rating, genres,
