+++ src/components/MovieDetailsPageStarRating/index.js
@@ -1,0 +1,3 @@
+import StarRating from './StarRating';
+
+export default StarRating;
