+++ src/components/MovieDetailsPageStarRating/starRating.scss
