+++ src/components/MovieListItem/MainInfo/index.js
