+++ src/components/MovieListItem/MainInfo/mainInfo.scss
