+++ src/components/MovieListItem/movieListItem.scss
