+++ src/components/MoviewList/index.js
@@ -1,0 +1,3 @@
+import MovieList from './MovieList';
+
+export default MovieList;
