+++ src/modules/module_Search/actions.js
@@ -11,9 +11,7 @@ export const fetchTrailer = (url) => {
 
 export const getFilms = () => {
   return dispatch => {
-   console.log('searching');
     const search = document.getElementById('searchInputID').value;
-    console.log('______________search', search);
     document.getElementById('searchInputID').value=''; 
     fetch(`https://api.themoviedb.org/3/search/movie?api_key=549663e4fb316b398fa37766692d00b7&language=en-US&query=${search}&page=1&include_adult=false`)
       .then(response => response.json()).then(data => dispatch(fetchData(data)));
@@ -32,10 +30,6 @@ export const getTrailer = (e) => {
   return dispatch => {
   fetch(`http://api.themoviedb.org/3/movie/${id}/videos?api_key=549663e4fb316b398fa37766692d00b7`)
     .then(response => response.json()).then(data => {
-      // const elem = document.getElementById(`trailerContainer${id}`);
-      // elem.classList.add('movieList__item__trailerContainerShow');
-      // console.log(`https://www.youtube.com/embed/${data.results[0].key}`);
-      // window.open(`https://www.youtube.com/embed/${data.results[0].key}`, 'trailer')
       return dispatch(fetchTrailer(`https://www.youtube.com/embed/${data.results[0].key}`));
     });
   }
